// src/app/interfaces/pokemon.ts
  export interface Species {
    name: string;
    url: string;
  }
  
  export interface DreamWorld {
    front_default: string;
    front_female?: any;
  }
  
  export interface Home {
    front_default: string;
    front_female?: any;
    front_shiny: string;
    front_shiny_female?: any;
  }
  
  export interface OfficialArtwork {
    front_default: string;
  }
  
  export interface Other {
    dream_world: DreamWorld;
    home: Home;
    "official-artwork": OfficialArtwork;
  }
// ...
  export enum Versions {
    "generation-i" = "GenerationI",
    "generation-ii" = "GenerationIi",
    "generation-iii" = "GenerationIii",
    "generation-iv" = "GenerationIv",
    "generation-v" = "GenerationV",
    "generation-vi" = "GenerationVi",
    "generation-vii" = "GenerationVii",
    "generation-viii" = "GenerationViii"
  }
  
  export interface Sprites {
    back_default: string;
    back_female?: any;
    back_shiny: string;
    back_shiny_female?: any;
    front_default: string;
    front_female?: any;
    front_shiny: string;
    front_shiny_female?: any;
    other: Other;
    versions: Versions;
  }
  
  export interface Stat {
    base_stat: number;
    effort: number;
    stat: Stat;
  }
  
  export interface Type {
    slot: number;
    type: TypeDetail;
  }
  
  export interface TypeDetail {
    name: string;
    url: string;
  }
  
  export interface Generation {
    name: string;
    url: string;
  }
  
  
  export interface PastType {
    generation: Generation;
    types: Type[];
  }
  
  export interface Pokemon {
    id?: string;
    name: string;
    base_experience: number;
    height: number;
    is_default: boolean;
    order: number;
    weight: number;
    location_area_encounters: string;
    sprites: Sprites;
    types: Type[];
    species: Species;
    /* 
    //NO SE USAN POR AHORA 
    abilities: Ability[];
    forms: Form[];
    game_indices: GameIndice[];
    held_items: HeldItem[];
    moves: Move[];
    stats: Stat[];
    past_types: PastType[];
    */
  }